requirements.txt
altgraph==0.17.2
anyio==3.5.0
asgiref==3.5.0
asyncio==3.4.3
blessed==1.19.0
blessings==1.7
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.4
codecov==2.1.12
colorama==0.4.4
coloredlogs==15.0.1
colour-runner==0.1.1
commonmark==0.9.1
coverage==6.2
cycler==0.11.0
deepdiff==5.7.0
distlib==0.3.4
fastapi==0.74.1
filelock==3.4.2
fonttools==4.28.5
h11==0.13.0
humanfriendly==10.0
idna==3.3
inquirer==2.9.1
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.1
ordered-set==4.0.2
packaging==21.3
pandas==1.3.5
pick==1.2.0
Pillow==9.0.0
platformdirs==2.4.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.9.0
Pygments==2.11.2
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.2
pyparsing==3.0.6
PySide6==6.2.2.1
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.3
readchar==3.0.5
requests==2.27.1
rich==11.0.0
rootpath==0.1.1
scipy==1.7.3
shiboken6==6.2.2.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
termcolor==1.1.0
toml==0.10.2
torch==1.9.0+cpu
torchaudio==0.9.0
torchvision==0.10.0+cpu
tox==3.24.5
typing-extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.5
virtualenv==20.13.0
wcwidth==0.2.5
